backend/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
google-generativeai
python-dotenv
pydantic
requests